fix(release): reject sidecar-only registry entries
