+++ scripts/verify_release_registry.py
@@ -11,67 +11,36 @@ import urllib.error
 import urllib.parse
 import urllib.request
 from collections.abc import Callable, Mapping, Sequence
-from dataclasses import dataclass
-from enum import Enum
 from pathlib import Path
 from typing import Final
 
 from packaging.utils import InvalidSdistFilename, InvalidWheelFilename, parse_sdist_filename, parse_wheel_filename
 from packaging.version import InvalidVersion, Version
 
+if __package__:
+    from .release_registry_types import (
+        Registry,
+        RegistryResult,
+        RegistryVerificationError,
+        ReleaseFile,
+        ReleaseInspection,
+        TestPyPIResult,
+    )
+else:
+    from release_registry_types import (
+        Registry,
+        RegistryResult,
+        RegistryVerificationError,
+        ReleaseFile,
+        ReleaseInspection,
+        TestPyPIResult,
+    )
+
 PROJECT_NAME: Final = "hol-guard"
 MAX_RESPONSE_BYTES: Final = 256 * 1024 * 1024
 _SHA256: Final[re.Pattern[str]] = re.compile(r"[0-9a-f]{64}")
 
 Fetcher = Callable[[str], bytes]
-
-
-class Registry(str, Enum):
-    PYPI = "pypi"
-    TESTPYPI = "testpypi"
-
-    @property
-    def api_host(self) -> str:
-        return "pypi.org" if self is Registry.PYPI else "test.pypi.org"
-
-    @property
-    def file_host(self) -> str:
-        return "files.pythonhosted.org" if self is Registry.PYPI else "test-files.pythonhosted.org"
-
-
-class RegistryVerificationError(RuntimeError):
-    pass
-
-
-@dataclass(frozen=True)
-class ReleaseFile:
-    filename: str
-    sha256: str
-    download_url: str
-
-
-@dataclass(frozen=True)
-class ReleaseInspection:
-    registry: Registry
-    version: str
-    exists: bool
-    files: tuple[ReleaseFile, ...] = ()
-
-    @property
-    def digests(self) -> dict[str, str]:
-        return {item.filename: item.sha256 for item in self.files}
-
-
-@dataclass(frozen=True)
-class RegistryResult:
-    registry: Registry
-    status: str
-    version: str
-    files: tuple[str, ...]
-    downloaded_paths: tuple[Path, ...] = ()
-
-
-TestPyPIResult = RegistryResult
 
 
 def stdlib_fetch(url: str) -> bytes:
